Base font size can be now easily changed in editor
There were bugs when changing themes, or setting themes overrides, new label settings... when the script was already added, so the user had to take out the script and re-add it. Now that's fixed. Whenever the user sets a new font size resource, the base font will be set accordingly.