change access to the manager autoload
The plugin couldn't be enabled since before enabling it the manager class wouldn't exist as an autoload, and that would create issues. While developing for G3, I fixed it by commenting out the references to that class and enable it, but didn't realize the problem after trying to enable it inside a clean project.

Initially I had the problem because in G4 you can have static variables (which I'm using) but in G3 those didn't exist yet. Hence the need for the singleton.